Added product piece service

+++ sephora-backend/Application/Services/Implementations/DeliveryService.cs
@@ -10,7 +10,7 @@ public class DeliveryService(
             await deliveryRepository.GetAll()
         );
 
-    public async Task<DeliveryDto?> GetById(int id)
+    public async Task<DeliveryDto?> GetById(long id)
     {
         DeliveryEntity? entity = await deliveryRepository.GetById(id);
         return entity is null ? null : mapper.Map<DeliveryDto>(entity);
@@ -30,7 +30,7 @@ public class DeliveryService(
         await deliveryRepository.Save();
     }
 
-    public async Task Delete(int id)
+    public async Task Delete(long id)
     {
         await deliveryRepository.Delete(id);
         await deliveryRepository.Save();
