+++ sephora-backend/Application/Services/Interfaces/IDeliveryService.cs
@@ -3,8 +3,8 @@ namespace CleanArchitecture.Application.Services.Interfaces;
 public interface IDeliveryService
 {
     Task<IEnumerable<DeliveryDto>> Get();
-    Task<DeliveryDto?> GetById(int id);
+    Task<DeliveryDto?> GetById(long id);
     Task Create(CreateDeliveryDto deliveryDto);
     Task Edit(DeliveryDto deliveryDto);
-    Task Delete(int id);
+    Task Delete(long id);
 }
